Refactor the form by extracting the CardForm out of the card page
