+++ lib/assets/styles/components/card-form.css
@@ -1,0 +1,16 @@
+.card-form {
+  width: 100%;
+  margin: 0 auto;
+  padding: 12px 24px;
+
+  @media (--breakpoint-small) {
+    width: 90%;
+    padding: 12px 0;
+  }
+
+  & .el-button {
+    width: 100%;
+    margin-left: 0;
+    margin-bottom: 12px;
+  }
+}
